boards: opentitan: Fixup existing test cases
