+++ boards/opentitan/src/tests/mod.rs
@@ -55,6 +55,7 @@ fn trivial_assertion() {
 
 mod aes_test;
 mod csrng;
+mod flash;
 mod hmac;
 mod multi_alarm;
 mod otbn;
@@ -63,9 +64,3 @@ mod rsa_4096;
 mod sha256soft_test; // Test software SHA capsule
 mod sip_hash;
 mod spi_host;
-mod tickv_test;
-/// Only run the flash_ctrl tests last, as testing memory protection
-/// may deny access to flash pages for other tests depending
-/// on flash (i.e tickV).
-#[rustfmt::skip]
-mod flash_ctrl;
